update input crate
avoids crashing on every shutdown
i love light entry so much!!

sidenote: how sending relative mouse movements as u32 ever work ???

+++ windows-gaming/driver/src/monitor/codec.rs
@@ -31,7 +31,7 @@ pub enum QmpCommand {
 pub enum InputEvent {
     Rel {
         axis: &'static str,
-        value: u32,
+        value: i64,
     },
     Btn {
         button: InputButton,
